# docs/api/snapshots/system-openapi.json
{
  "components": {
    "schemas": {
      "CreateItemRequest": {
        "properties": {
          "description": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "key": {
            "type": "string"
          },
          "sortOrder": {
            "format": "int32",
            "type": "integer"
          },
          "value": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "CreateTypeRequest": {
        "properties": {
          "code": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ItemView": {
        "properties": {
          "description": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "sortOrder": {
            "format": "int32",
            "type": "integer"
          },
          "typeId": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "LocaleView": {
        "properties": {
          "defaultLocale": {
            "type": "boolean"
          },
          "displayName": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "localeCode": {
            "type": "string"
          },
          "nativeName": {
            "type": "string"
          },
          "sortOrder": {
            "format": "int32",
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "ResolvedSystemDictionaryItem": {
        "properties": {
          "dictionaryCode": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "selectable": {
            "type": "boolean"
          },
          "value": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ResultItemView": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ItemView"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ResultListItemView": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/ItemView"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ResultListSystemDictionaryItemOption": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/SystemDictionaryItemOption"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ResultListTypeView": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/TypeView"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ResultLocaleView": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/LocaleView"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ResultResolvedSystemDictionaryItem": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ResolvedSystemDictionaryItem"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ResultTypeView": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/TypeView"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ResultVoid": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {},
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "StatusRequest": {
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "SystemDictionaryItemOption": {
        "properties": {
          "dictionaryCode": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "sortOrder": {
            "format": "int32",
            "type": "integer"
          },
          "value": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "TypeView": {
        "properties": {
          "code": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "UpdateItemRequest": {
        "properties": {
          "description": {
            "type": "string"
          },
          "sortOrder": {
            "format": "int32",
            "type": "integer"
          },
          "value": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "UpdateTypeRequest": {
        "properties": {
          "description": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "服务内部 Controller 路径；Gateway 路径与鉴权规则见 docs/api/README.md。",
    "title": "MOM System Service API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/admin/dictionaries": {
      "get": {
        "operationId": "listTypes",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListTypeView"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "tags": [
          "dictionary-controller"
        ]
      },
      "post": {
        "operationId": "createType",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTypeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultTypeView"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "tags": [
          "dictionary-controller"
        ]
      }
    },
    "/admin/dictionaries/{id}": {
      "put": {
        "operationId": "updateType",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTypeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultTypeView"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "tags": [
          "dictionary-controller"
        ]
      }
    },
    "/admin/dictionaries/{typeId}/items": {
      "get": {
        "operationId": "listItems",
        "parameters": [
          {
            "in": "path",
            "name": "typeId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListItemView"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "tags": [
          "dictionary-controller"
        ]
      },
      "post": {
        "operationId": "createItem",
        "parameters": [
          {
            "in": "path",
            "name": "typeId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultItemView"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "tags": [
          "dictionary-controller"
        ]
      }
    },
    "/admin/dictionaries/{typeId}/items/{itemId}": {
      "put": {
        "operationId": "updateItem",
        "parameters": [
          {
            "in": "path",
            "name": "typeId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "itemId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultItemView"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "tags": [
          "dictionary-controller"
        ]
      }
    },
    "/admin/dictionaries/{typeId}/items/{itemId}/status": {
      "patch": {
        "operationId": "changeItemStatus",
        "parameters": [
          {
            "in": "path",
            "name": "typeId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "itemId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultItemView"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "tags": [
          "dictionary-controller"
        ]
      }
    },
...
    "/dictionaries/{dictionaryCode}/items": {
      "get": {
        "operationId": "activeItems",
        "parameters": [
          {
            "in": "path",
            "name": "dictionaryCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListSystemDictionaryItemOption"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "tags": [
          "dictionary-controller"
        ]
      }
    },
    "/dictionaries/{dictionaryCode}/items/{key}": {
      "get": {
        "operationId": "resolveItem",
        "parameters": [
          {
            "in": "path",
            "name": "dictionaryCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultResolvedSystemDictionaryItem"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "tags": [
          "dictionary-controller"
        ]
      }
    },
...
    "/i18n/locales/admin/{id}/status": {
      "patch": {
        "operationId": "changeStatus",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultLocaleView"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "tags": [
          "supported-locale-controller"
        ]
      }
    },
...
    "/{id}/status": {
      "patch": {
        "operationId": "changeTypeStatus",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultTypeView"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "tags": [
          "dictionary-controller"
        ]
      }
    }
  }
}